# Make FVA a bit faster.