The FVA in cobrapy alternated between optimizing via minimum and maximum, but the minimum solutions tend to be closer to each other than to the maximum solutions and vice versa. Thus, a better recycling strategy for the solution basis is to minimize all reactions first followed by maximizing them. This is also the algorithm described in the original publication.

## Benchmarks

Recon 2.2, more than 7500 reactions

```python
from cobra.flux_analysis import flux_variability_analysis

%time va = flux_variability_analysis(recon2)
%time va = flux_variability_analysis(recon2, solver="cplex")
%time va = flux_variability_analysis(recon2, solver="gurobi")
```

Old:

```
cglpk:
CPU times: user 2min 5s, sys: 4 ms, total: 2min 5s
Wall time: 2min 5s

cplex:
CPU times: user 59.4 s, sys: 4 ms, total: 59.4 s
Wall time: 59.4 s

gurobi:
CPU times: user 1min 28s, sys: 4 ms, total: 1min 28s
Wall time: 1min 28s

```

New:

```
cglpk:
CPU times: user 1min 30s, sys: 8 ms, total: 1min 30s
Wall time: 1min 30s

cplex:
CPU times: user 47.2 s, sys: 16 ms, total: 47.2 s
Wall time: 47.2 s

gurobi:
CPU times: user 1min 16s, sys: 24 ms, total: 1min 16s
Wall time: 1min 16s
```

More or less 25% speed up, mostly for cglpk which is a bit better in recycling previous solutions in general.