@cdiener 2016-10-07
The basis is part of the LP object, named `lp` in the code here. The trick is to declare one LP object and solve it with different coefficients and not use model functions and `optimize`. This is what happens on lines 59-60. Also see [chapter 11 of the docs](https://cobrapy.readthedocs.io/en/latest/solvers.html).