@marouenbg 2016-10-07
Thanks. I am a new user of cobrapy and when comparing to MATLAB where you have to specifically set the solver to output the solution basis, I couldn't find in the different functions any advanced start parameter turned on. Maybe this is done automatically.