@cdiener 2016-10-07
Luckily in Python we have objects. So we don't have to pass around the state all the time. The basis is kept within the LP objects after solving the first time and is automatically used the second time you solve it.